requirements.txt
click
python-dotenv
pyyaml
paramiko